Customize the Hello View
There is default code added to Views that are generated using the
`rails` command. There is a page heading which contains text in the
format ControllerName#viewname (in our example, Say#hello). This
heading is followed by a paragraph which explains where to find the
view on the file system (in our example it says "Find me in
app/views/say/hello.html.erb")

In this exercise, we modify the heading so that it says "Hello from
Rails!" and we add a paragraph tag that includes the text:

It is now <%= Time.now %>

The <%= %> tags are used to output the result of an embedded Ruby
code statement (similar to using puts or print commands in the
console.