Refactor the Hello View to Call an Instance Variable from the Controller
This example introduces instance variables. Our SayController is a Ruby
class that inherits from ApplicationController. When we generated the
controller using the `rails generate` command, we passed in two views:
hello and goodbye. Rails automatically added two methods to the class:

def hello
end

def goodbye
end

We move the code from the view that outputs the time (Time.now) into the
hello method, and assign it to the variable @time. The variable can be
called from within the hello view.

We refactor the view so that the embedded Ruby statement reads:

<%= @time %>

On the surface, this might appear more verbose than the original version.
It requires more code to implement, and we do use more characters. This is
not necessarily a bad thing. This honors the "separation between presentation
and content" that we strive for when writing markup, and it provides a
reusable implementation of the current time as a variable.