Add Links to Navigate Between Views
In this exercise, we use the Rails link_to helper method to add
anchor tags to the page that provide links from the Hello to the
Goodbye view and vice versa.

The link_to method takes two parameters: the link text and the
href, or path that the link should navigate to. In Rails, method
parameters can be wrapped in parentheses as in other languages,
but this is not required.

* app/views/say/goodbye.html.erb
* app/views/say/hello.html.erb